openmaptiles2mimir: spread the load in the thread pool (#86)
`load_and_index_pois` was generaly not thread safe so most of the
computation was done in a single thread which appears to be blocking.

Sadly, a big part of the change mostly addresses lifetime issues because
of `tokio::spawn` requiring a 'static future. In our case something like
tokio-scoped could lift this requirement but the very few clones and
`Arc`' wouldn't justify a new dependancy.