some cosmetic cleanup
I have the intention to showcase Rust based on this repository, so I've
tryied to remove some dust while reading through the code. The aim is
mostly to make it more readable, straight forward (I would love to add
beginner friendly, but fafnir is far from being simple, especially for
what it does).

The diff is poluted with format string updates but there are also some
more refined changes, such as sparing an async block in
`parse_properties`.