tripadvisor: use serde to stream over the feeds
The file `parser.rs` got even more complicated with the support of new
JSON feeds while it seems that reading JSON is easier and has better
tooling.

I was able to build a similar parser over serde. This means that we are
now robust to any kind of formatting, regardless of the feed format. The
con is that serde doesn't support async, so I had to spawn a dedicated
thread to handle this.